Refactored bot into seperate modules
